requirements.txt
anytree
appdirs
aria2p
attrdict
beautifulsoup4
bencoding
cfscrape
feedparser
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
pillow
psutil
psycopg2-binary
pybase64
pyrogram
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
six
telegraph
tenacity
tgCrypto
urllib3
yt-dlp
speedtest-cli
cloudscrape
torrentool
heroku3
js2py
aiohttp
gitpython